update: support randomize env
