+++ scheduler/scheduler.go
@@ -5,19 +5,18 @@ import (
 	"io"
 	"net/http"
 	"os"
-	"strconv"
-	"strings"
 	"time"
 
 	"github.com/go-co-op/gocron/v2"
 	"github.com/ishikawa096/auto-requester/utils"
 )
 
-type httpOptions struct {
+type requestOptions struct {
 	method      string // GET, POST, PUT, DELETE
 	url         string
 	contentType string
 	body        []byte
+	randomize   bool
 }
 
 var Scheduler gocron.Scheduler
@@ -32,9 +31,16 @@ func init() {
 	}
 }
 
-func sendRequest(options httpOptions) (*http.Response, error) {
-	// リクエストボディを再利用可能な形にするために、bytes.Readerを使用
-	bodyReader := bytes.NewReader(options.body)
+func sendRequest(options requestOptions) (*http.Response, error) {
+	var requestBody []byte = nil
+	if options.randomize {
+		requestBody = selectRandomElement(options.body)
+	} else {
+		requestBody = options.body
+	}
+
+	// To use the request body multiple times, we need to create a reader
+	bodyReader := bytes.NewReader(requestBody)
 
 	req, err := http.NewRequest(options.method, options.url, bodyReader)
 	if err != nil {
@@ -51,9 +57,9 @@ func sendRequest(options httpOptions) (*http.Response, error) {
 	return resp, nil
 }
 
-func createTask(options httpOptions) gocron.Task {
+func createTask(options requestOptions) gocron.Task {
 	return gocron.NewTask(
-		func(options httpOptions) {
+		func(options requestOptions) {
 			resp, err := sendRequest(options)
 			if err != nil {
 				utils.Logger("Error sending request:", err)
@@ -72,7 +78,7 @@ func createTask(options httpOptions) gocron.Task {
 	)
 }
 
-func registerJob(minSec, maxSec int, options httpOptions) {
+func registerJob(minSec, maxSec int, options requestOptions) {
 	_, err := Scheduler.NewJob(
 		gocron.DurationRandomJob(
 			time.Duration(minSec)*time.Second,
@@ -86,66 +92,8 @@ func registerJob(minSec, maxSec int, options httpOptions) {
 	}
 }
 
-func getRequestBody() []byte {
-	file, err := os.Open("/etc/app/body.json")
-	if err != nil {
-		utils.Logger("Error opening file:", err)
-		os.Exit(1)
-	}
-	defer file.Close()
-	// read the file content to memory
-	fileContent, err := io.ReadAll(file)
-	if err != nil {
-		utils.Logger("Error reading file:", err)
-		os.Exit(1)
-	}
-	return fileContent
-}
-
-func getSettingValues() (int, int, httpOptions) {
-	// default values
-	minSec := 3
-	maxSec := 5
-	url := "http://localhost:3000"
-	method := "GET"
-	contentType := "application/json"
-
-	if val, exists := os.LookupEnv("INTERVAL_MIN_SEC"); exists {
-		if v, err := strconv.Atoi(val); err == nil {
-			minSec = v
-		}
-	}
-	if val, exists := os.LookupEnv("INTERVAL_MAX_SEC"); exists {
-		if v, err := strconv.Atoi(val); err == nil {
-			maxSec = v
-		}
-	}
-	if val, exists := os.LookupEnv("HTTP_METHOD"); exists {
-		method = strings.ToUpper(val)
-	}
-	if val, exists := os.LookupEnv("TARGET_URL"); exists {
-		url = val
-	}
-	if val, exists := os.LookupEnv("CONTENT_TYPE"); exists {
-		contentType = val
-	}
-
-	var requestBody []byte = nil
-	if method == "POST" || method == "PUT" {
-		requestBody = getRequestBody()
-	}
-
-	options := httpOptions{
-		method:      method,
-		url:         url,
-		contentType: contentType,
-		body:        requestBody,
-	}
-	return minSec, maxSec, options
-}
-
 func StartJob() {
-	registerJob(getSettingValues())
+	registerJob(getConfigs())
 
 	// start the scheduler
 	Scheduler.Start()
